chore!: modernize the toolchain and rename to the ai.openspace package
Foundation for the v8 rewrite. No behavior changes — every Kotlin source is
byte-identical to master apart from its package declaration.

Dependencies move to what Diana ships: React Native 0.84.1, React 19.2.3,
TypeScript 5.9. The example app moves with them (Gradle 8.14.3, Kotlin 2.1.20,
SDK 36, newArchEnabled) so CI exercises the same toolchain a consumer gets.

Android switches to the React Native Gradle plugin, which supplies the
react-android version and will run codegen once a spec exists. The library now
inherits compileSdk/minSdk/targetSdk from the host app instead of pinning them,
with fallbacks matching Diana. The manifest `package` attribute moves to
`android.namespace`, as AGP 8 requires.

The `com.vydia.RNUploader` package becomes `ai.openspace.backgroundupload`.
This is a source-level rename only; the JS module name is unchanged, so nothing
on the JS side moves.

Packaging: `files` allowlist so publishing stops shipping internal notes and
build output, `typings` points at source, and the checked-in `lib/*.d.ts` build
output is deleted. CI gains typecheck and Android unit-test steps.

BREAKING CHANGE: requires React Native 0.84+ and React 19+. Consumers with
ProGuard rules referencing com.vydia.RNUploader must update the package name.
