+++ .github/workflows/node.yml
@@ -1,26 +1,39 @@
-name: Node Environment
+name: CI
 
 on: [push, pull_request]
 
 jobs:
-  node-lint-tests:
+  checks:
     runs-on: ubuntu-latest
     if: "!contains(github.event.head_commit.message, '[skip ci]')"
 
     steps:
-    - name: checkout
-      uses: actions/checkout@v2
+      - name: checkout
+        uses: actions/checkout@v4
 
-    - name: setup node
-      uses: actions/setup-node@v1
-      with:
-        node-version: 20
+      - name: setup node
+        uses: actions/setup-node@v4
+        with:
+          node-version: 20
 
-    - name: install node_modules
-      run: | 
-        yarn install --frozen-lockfile
-        yarn --cwd example/RNBGUExample install --frozen-lockfile 
+      - name: install node_modules
+        run: |
+          yarn install --frozen-lockfile
+          yarn --cwd example/RNBGUExample install --frozen-lockfile
 
-    - name: node lint
-      run:
-        yarn lint:ci
+      - name: lint
+        run: yarn lint:ci
+
+      - name: typecheck
+        run: yarn typecheck
+
+      - name: setup java
+        uses: actions/setup-java@v4
+        with:
+          distribution: temurin
+          java-version: 17
+
+      - name: android unit tests
+        run: |
+          cd example/RNBGUExample/android
+          ./gradlew :react-native-background-upload:testDebugUnitTest
