+++ CONTRIBUTING.md
@@ -1,8 +1,5 @@
 ## Issues
 
-Along with a bug report, provide a functional example repository which
-reproduces the bug you're experiencing. We've made this very easy by providing
-a full-stack (React Native + Express.js) example app, which you can fork and
-alter to reproduce your bug:
-
-[ReactNativeBackgroundUploadExample](https://github.com/Vydia/ReactNativeBackgroundUploadExample)
+Along with a bug report, provide a functional reproduction using the full-stack
+(React Native + Express.js) example app in this repo under `example/` — fork and
+alter it to reproduce your bug.
