+++ android/src/main/java/ai/openspace/backgroundupload/EventReporter.kt
@@ -1,4 +1,4 @@
-package com.vydia.RNUploader
+package ai.openspace.backgroundupload
 
 import android.util.Log
 import com.facebook.react.bridge.Arguments
