+++ example/RNBGUExample/package.json
@@ -12,8 +12,8 @@
   "dependencies": {
     "@notifee/react-native": "^9.1.8",
     "path": "^0.12.7",
-    "react": "18.3.1",
-    "react-native": "0.75.3",
+    "react": "19.2.3",
+    "react-native": "0.84.1",
     "react-native-fs": "^2.20.0"
   },
   "devDependencies": {
@@ -21,18 +21,17 @@
     "@babel/preset-env": "^7.20.0",
     "@babel/runtime": "^7.20.0",
     "@react-native-community/cli": "^14.1.1",
-    "@react-native/babel-preset": "0.75.3",
-    "@react-native/eslint-config": "0.75.3",
-    "@react-native/metro-config": "0.75.3",
-    "@react-native/typescript-config": "0.75.3",
-    "@types/react": "^18.2.6",
-    "@types/react-test-renderer": "^18.0.0",
+    "@react-native/babel-preset": "0.84.1",
+    "@react-native/eslint-config": "0.84.1",
+    "@react-native/metro-config": "0.84.1",
+    "@react-native/typescript-config": "0.84.1",
+    "@types/react": "^19.2.15",
     "babel-jest": "^29.6.3",
     "babel-plugin-module-resolver": "^5.0.2",
     "eslint": "^8.19.0",
     "jest": "^29.6.3",
     "prettier": "2.8.8",
-    "react-test-renderer": "18.3.1",
+    "react-test-renderer": "19.2.3",
     "typescript": "^5.9.3"
   },
   "engines": {
