--- lib/index.d.ts
@@ -1,28 +1,0 @@
-import { AddListener, UploadOptions } from './types';
-export * from './types';
-declare const _default: {
-    startUpload: ({ path, android, ios, ...options }: UploadOptions) => Promise<string>;
-    cancelUpload: (cancelUploadId: string) => Promise<boolean>;
-    addListener: AddListener;
-    ios: {
-        /**
-         * Directly check the state of a single upload task without using event listeners.
-         * Note that this method has no way of distinguishing between a task being completed, errored, or non-existent.
-         * They're all `undefined`. You will need to either rely on the listeners or
-         * check with the API service you're using to upload.
-         */
-        getUploadStatus: (jobId: string) => Promise<{
-            state: "running" | "suspended" | "canceling";
-            bytesSent: number;
-            totalBytes: number;
-        } | undefined>;
-    };
-    android: {
-        /**
-         * When the upload progress notification is pressed, it will open the app and fire this event
-         * @param listener
-         */
-        addNotificationListener: (listener: () => void) => import("react-native").EmitterSubscription;
-    };
-};
-export default _default;
