+++ package.json
@@ -1,13 +1,38 @@
 {
   "name": "react-native-background-upload",
-  "version": "7.5.3",
+  "version": "8.0.0",
   "description": "Cross platform http post file uploader with android and iOS background support",
   "main": "src/index",
-  "typings": "lib/index.d.ts",
+  "typings": "src/index.ts",
   "react-native": "src/index",
   "source": "src/index",
+  "files": [
+    "src",
+    "!src/__tests__",
+    "android",
+    "!android/build",
+    "!android/.gradle",
+    "ios",
+    "react-native-background-upload.podspec",
+    "README.md",
+    "CHANGELOG.md",
+    "LICENSE"
+  ],
+  "codegenConfig": {
+    "name": "RNFileUploaderSpec",
+    "type": "modules",
+    "jsSrcsDir": "src",
+    "android": {
+      "javaPackageName": "ai.openspace.backgroundupload"
+    },
+    "ios": {
+      "modulesProvider": {
+        "RNFileUploader": "RNFileUploader"
+      }
+    }
+  },
   "scripts": {
-    "build": "tsc && tsc-alias",
+    "typecheck": "tsc",
     "lint": "yarn lint-root --fix && yarn lint-example --fix",
     "lint:ci": "yarn lint-root && yarn lint-example",
     "lint-root": "eslint --ext js,jsx,ts,tsx src --quiet",
@@ -15,7 +40,7 @@
   },
   "repository": {
     "type": "git",
-    "url": "https://github.com/Vydia/react-native-background-upload.git"
+    "url": "https://github.com/openspacelabs/react-native-background-upload.git"
   },
   "keywords": [
     "NSURLSession",
@@ -26,25 +51,25 @@
     "react"
   ],
   "peerDependencies": {
-    "react": "*",
-    "react-native": ">=0.47.0"
+    "react": ">=19.0.0",
+    "react-native": ">=0.84.0"
   },
-  "author": "Steve Potter",
-  "license": "BSD-3-Clause",
+  "author": "OpenSpace",
+  "license": "MIT",
   "bugs": {
-    "url": "https://github.com/Vydia/react-native-background-upload/issues"
+    "url": "https://github.com/openspacelabs/react-native-background-upload/issues"
   },
-  "homepage": "https://github.com/Vydia/react-native-background-upload#readme",
+  "homepage": "https://github.com/openspacelabs/react-native-background-upload#readme",
   "devDependencies": {
     "@babel/core": "^7.20.0",
     "@babel/preset-env": "^7.20.0",
+    "@babel/preset-typescript": "^7.20.0",
     "@babel/runtime": "^7.20.0",
-    "@react-native/babel-preset": "0.75.3",
-    "@react-native/eslint-config": "0.75.3",
-    "@react-native/metro-config": "0.75.3",
-    "@react-native/typescript-config": "0.75.3",
-    "@types/react": "^18.2.6",
-    "@types/react-test-renderer": "^18.0.0",
+    "@react-native/babel-preset": "0.84.1",
+    "@react-native/eslint-config": "0.84.1",
+    "@react-native/metro-config": "0.84.1",
+    "@react-native/typescript-config": "0.84.1",
+    "@types/react": "^19.2.15",
     "babel-jest": "^29.6.3",
     "eslint": "^8.19.0",
     "eslint-plugin-unused-imports": "^4.1.4",
@@ -52,12 +77,10 @@
     "jest": "^29.6.3",
     "lint-staged": "^10.0.8",
     "prettier": "2.8.8",
-    "react": "18.3.1",
-    "react-native": "0.75.3",
-    "react-test-renderer": "18.3.1",
+    "react": "19.2.3",
+    "react-native": "0.84.1",
     "semantic-release": "^15.13.31",
-    "tsc-alias": "^1.6.8",
-    "typescript": "5.0.4"
+    "typescript": "^5.9.3"
   },
   "release": {
     "plugins": [
